Removes requirement for a proxy-object.
Changes how code generation works, removing the need to construct an object-literal to act as a proxy between the imported values and their usage.
This extends to object-destructuring as well, which handles many cases more safely now.
The `xforms` object defines the behaviors to transform from the usage of a namespace import to a named import.
Further simplifies the `ImportDeclaration` visitor, extracting a lot of its work into the `UidMap` class, which manages the mapping between pre-existing identifiers and their aliases and can also produce the specifiers for those mappings.
The `engine` requirement for `node` has been bumped up to >= version 6, due to the use of certain patterns related to `const` and `let`.