Now visits `ImportNamespaceSpecifier` instead.
Simplifies the plugin's logic; reduces usage of `Function#bind`, which has some overhead.
Separates a transformed specifier into its own `ImportDeclaration`, which allows other plugins the opportunity to perform additional transforms on the imports.
Adds a test for combining a wildcard import with a default import, which is valid syntax.