Fixes a possible issue where an incomplete transform was possible.
The `tryDoTransforms` function is supposed to abort if it comes across a statement it isn't equipped to process, however, it was possible that it could perform some transformations before aborting, leaving the AST in an incorrect state.

The function was adjusted to only begin performing the transformations if every statement encountered is supported.